+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 [Downloadable releases](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
+- core hide tooltip on popup open
 - plugin admin add friend in user admin
 - plugin admin change place owner
 - new placeholder itemPlaceSearch